fix parsing json to serialization
