+++ app/build.gradle
@@ -53,7 +53,6 @@ dependencies {
 
     def lifecycleVersion = '2.4.0'
 
-
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycleVersion"
     implementation "androidx.fragment:fragment-ktx:1.4.1"
 
@@ -62,14 +61,12 @@ dependencies {
 
     implementation "androidx.navigation:navigation-fragment-ktx:2.4.0"
 
-
     def moshiVersion = '1.13.0'
     implementation "com.squareup.moshi:moshi:$moshiVersion"
     implementation "com.squareup.moshi:moshi-kotlin:$moshiVersion"
 
     implementation "androidx.room:room-runtime:2.3.0"
     annotationProcessor "androidx.room:room-compiler:2.3.0"
-
 
     def retrofitVersion = '2.9.0'
     implementation "com.squareup.retrofit2:retrofit:$retrofitVersion"
@@ -79,7 +76,6 @@ dependencies {
     def coroutinesVersion = "1.5.1"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutinesVersion"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$coroutinesVersion"
-
 
     implementation 'net.openid:appauth:0.9.1'
     implementation 'com.jakewharton.timber:timber:5.0.1'
