+++ app/src/main/java/com/gitficko/github/remote/ApiClient.kt
@@ -1,8 +1,9 @@
-package company.vk.polis.github_mobile.remote
+package com.gitficko.github.remote
 
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import kotlinx.serialization.json.Json
 import okhttp3.MediaType
+import okhttp3.MediaType.Companion.toMediaType
 import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 
@@ -20,7 +21,7 @@ object ApiClient {
 
     private val retrofit: Retrofit = Retrofit.Builder()
         .baseUrl(BASE_URL)
-        .addConverterFactory(jsonConf.asConverterFactory(MediaType.get("application/json")))
+        .addConverterFactory(jsonConf.asConverterFactory("application/json".toMediaType()))
         .client(okHttpClient)
         .build()
