+++ app/src/main/java/com/gitficko/github/ui/dashboard/DashboardViewModel.kt
@@ -1,13 +1,13 @@
-package company.vk.polis.github_mobile.ui.notifications
+package com.gitficko.github.ui.dashboard
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class DashboardViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is notifications Fragment"
+        value = "This is dashboard Fragment"
     }
     val text: LiveData<String> = _text
 }
