+++ app/src/main/java/com/gitficko/github/ui/home/HomeFragment.kt
@@ -1,4 +1,4 @@
-package company.vk.polis.github_mobile.ui.home
+package com.gitficko.github.ui.home
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import company.vk.polis.github_mobile.databinding.FragmentHomeBinding
+import com.gitficko.github.databinding.FragmentHomeBinding
 
 class HomeFragment : Fragment() {
