+++ app/src/main/java/com/gitficko/github/ui/notifications/NotificationsViewModel.kt
@@ -1,13 +1,13 @@
-package company.vk.polis.github_mobile.ui.dashboard
+package com.gitficko.github.ui.notifications
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class NotificationsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is dashboard Fragment"
+        value = "This is notifications Fragment"
     }
     val text: LiveData<String> = _text
 }
